ai_lip_sync/requirements.txt
gTTS==2.5.4
mysql_connector_repackaged==0.3.1
numpy==2.2.4
protobuf==6.30.2
PyAudio==0.2.14
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.44.1
torch==2.6.0
transformers==4.51.1
